Fixed typos and created conftest.py in tests
