+++ tests/clustering/kmeans/test_distances.py
@@ -6,17 +6,6 @@ from directional_clustering.clustering import distance_cosine
 
 
 # ==============================================================================
-# Fixtures
-# ==============================================================================
-
-@pytest.fixture
-def random_array():
-    """
-    A 2x2 array with random float values.
-    """
-    return np.random.rand(2, 2)
-
-# ==============================================================================
 # Tests
 # ==============================================================================
